+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - fixed a bug where comparing `ExtensionLocalValue` would fail when using the `!=` operator [#304](https://github.com/Draegerwerk/sdc11073/issues/304)
 - fixed missing xml header in discovery messages [#314](https://github.com/Draegerwerk/sdc11073/issues/314)
-- fixed possible error in alarmprovider [#293](https://github.com/Draegerwerk/sdc11073/issues/293)
+- fixed possible error in alarmprovider (caching of handles outside transaction context) [#293](https://github.com/Draegerwerk/sdc11073/issues/293)
 - fixed a bug where the reason for a subscribe failure would not be logged correctly [#323](https://github.com/Draegerwerk/sdc11073/issues/323)
 - fixed a bug where `isConnected` may be `True`, even though the subscription is not subscribed, because `allSubscriptionsOkay` defaults value is `True` [#326](https://github.com/Draegerwerk/sdc11073/issues/326) and it needs one second bevor changing its value to `False`
 - `RuntimeError: dictionary changed size during iteration` for `remoteServices` in the discovery [#335](https://github.com/Draegerwerk/sdc11073/issues/335)